requirements.txt
transformers
huggingface_hub
streamlit
langchain_core
langchain_community
langchain_huggingface
langchain_text_splitters
accelerate
watchdog
tqdm
pyttsx3
openai-whisper
sounddevice
numpy
librosa
python-dotenv